Fix MongoState.delete(Boolean)
MongoState did not override AbstractState.destroy(Boolean). In AbstractState.destroy(Boolean) is the variable 'service' used, which is/can become null. The MongoState class uses primarily the 'provider' variable. This caused agents to be deleted from the InstantiationService, but not from the database.

Note: this should maybe be fixed in another way by calling getService() instead of using the 'provider' variable?